Refactored to have class based scrapers
