--- src/scraper/full_title_scraper.py
@@ -1,88 +1,0 @@
-import pandas as pd
-import requests
-
-from bs4 import BeautifulSoup
-
-
-def get_suburbs():
-    # Noted that this URL does change to SearchCrit2 for sectional title
-    # But assuming suburbs are the same for both full title and sectional title
-    url = 'http://valuation2022.durban.gov.za/FramePages/SearchCrit.aspx'
-    response = requests.get(url)
-    soup = BeautifulSoup(response.content, 'html.parser')
-
-    # Find the select tag by ID (optional)
-    select_tag = soup.find('select', id='drpSuburb')
-
-    # Find all option tags within the select tag
-    options = select_tag.find_all('option')
-
-    # Extract text and value for each option
-    option_data = []
-    for option in options:
-        text = option.text.strip()
-        value = option.get('value')
-
-        if value:
-            option_data.append({'text': text, 'value': value})
-
-    return pd.DataFrame(option_data)
-
-
-def process_cells(cells):
-    row_data = []
-    for cell in cells:
-        row_data.append(cell.text.strip())
-    return row_data
-
-
-def save_data(table):
-    data = []
-    headers = []
-
-    # Skip the first row which has the number of results
-    rows = table.find_all('tr')[1:]
-
-    for row in rows:
-        cells = row.find_all(['th', 'td'])
-        if not headers:
-            headers = process_cells(cells)
-        else:
-            data.append(process_cells(cells))
-
-    df = pd.DataFrame(data, columns=headers)
-    print(df)
-    # TODO: Actually save the data
-
-
-def scrape_full_title():
-    suburbs = get_suburbs()
-
-    url = 'http://valuation2022.durban.gov.za/FramePages/SearchResult.aspx'
-
-    # For each suburb, make request to get data
-    for index, suburb in suburbs.iterrows():
-        suburb_value = suburb["value"]
-        # TODO change roll based on whether full title or sectional title
-        search_url = (f"{url}?Roll=1&VolumeNo=&RateNumber=&StreetNo=&StreetName=&Suburb={suburb_value}"
-                      f"&ERF=&Portion=&DeedsTown=&SchemeName=&SectionNumber=&All=true")
-
-        print(search_url)
-        response = requests.get(search_url)
-        soup = BeautifulSoup(response.content, 'html.parser')
-        table = soup.find('table', class_='searchResultTable')
-
-        if table:
-            print("Table found")
-            save_data(table)
-
-        else:
-            print("Table not found")
-        break
-
-
-scrape_full_title()
-# Full title
-# "https://valuation2022.durban.gov.za/FramePages/SearchResult.aspx?Roll=1&VolumeNo=&RateNumber=&StreetNo=&StreetName=&Suburb=22136&ERF=&Portion=&DeedsTown=&SchemeName=&SectionNumber=&All=false"
-# Sectional title
-# "https://valuation2022.durban.gov.za/FramePages/SearchResult.aspx?Roll=2&VolumeNo=&RateNumber=&StreetNo=&StreetName=&Suburb=22136&ERF=&Portion=&DeedsTown=&SchemeName=&SectionNumber=&All=false"
